+++ tests/lantern_tests/models/record/test_revision.py
@@ -59,14 +59,14 @@ class TestRecordRevision:
             )
 
     @pytest.mark.parametrize(("extra_config", "expected"), [({}, True), ({"x": "x"}, False)])
-    def test_config_supported(self, fx_record_config_minimal_iso: dict, extra_config: dict, expected: bool):
+    def test_config_supported(self, fx_revision_config_min: dict, extra_config: dict, expected: bool):
         """Can accurately determine if a record config contains unsupported properties."""
-        config = {**fx_record_config_minimal_iso, "file_revision": "x", **extra_config}
+        config = {**fx_revision_config_min, **extra_config}
         result = RecordRevision._config_supported(config=config)
         assert result == expected
 
     @pytest.mark.parametrize("has_schema", [False, True])
-    def test_loads(self, has_schema: bool):
+    def test_loads(self, fx_revision_config_min: dict, has_schema: bool):
         """
         Can create a Record Revision from a JSON serialised dict plus additional context.
 
@@ -75,23 +75,13 @@ class TestRecordRevision:
         expected_str = "x"
         schema = "https://metadata-resources.data.bas.ac.uk/bas-metadata-generator-configuration-schemas/v2/iso-19115-2-v4.json"
 
-        base_config = {
-            "hierarchy_level": HierarchyLevelCode.DATASET,
-            "metadata": {
-                "contacts": [{"organisation": {"name": expected_str}, "role": [ContactRoleCode.POINT_OF_CONTACT]}],
-                "date_stamp": date(2014, 6, 30).isoformat(),
-            },
-            "identification": {
-                "title": {"value": expected_str},
-                "dates": {"creation": date(2014, 6, 30).isoformat()},
-                "abstract": expected_str,
-            },
-        }
-        config = {**base_config, "file_revision": expected_str}
+        fx_revision_config_min["file_identifier"] = expected_str
+        fx_revision_config_min["file_revision"] = expected_str
+        fx_revision_config_min["identification"]["title"]["value"] = expected_str
         if has_schema:
-            config["$schema"] = schema
+            fx_revision_config_min["$schema"] = schema
 
-        record = RecordRevision.loads(config)
+        record = RecordRevision.loads(fx_revision_config_min)
 
         assert record.identification.title == expected_str  # record property
         assert record.file_revision == expected_str  # record revision property
@@ -104,20 +94,15 @@ class TestRecordRevision:
             _ = RecordRevision.loads(fx_revision_config_min)
 
     @pytest.mark.parametrize("inc_revision", [False, True])
-    def test_dumps(
-        self, fx_record_config_minimal_iso: dict, fx_record_revision_minimal_iso: RecordRevision, inc_revision: bool
-    ):
+    def test_dumps(self, fx_revision_config_min: dict, fx_revision_model_min: RecordRevision, inc_revision: bool):
         """
         Can encode record revision as a dict that can be serialised to JSON with optional file revision property.
 
         This only tests revision specific properties can be optionally included along with regular Record properties.
         """
-        config = fx_record_revision_minimal_iso.dumps(with_revision=inc_revision)
+        config = fx_revision_model_min.dumps(with_revision=inc_revision)
 
-        assert (
-            config["identification"]["title"]["value"]
-            == fx_record_config_minimal_iso["identification"]["title"]["value"]
-        )
+        assert config["identification"]["title"]["value"] == fx_revision_config_min["identification"]["title"]["value"]
         if inc_revision:
             assert config["file_revision"] == "x"
         else: