+++ tests/lantern_tests/templates/test_catalogue_item.py
@@ -8,14 +8,14 @@ from lantern.models.item.catalogue import ItemCatalogue
 class TestItemTemplate:
     """Test base catalogue item template."""
 
-    def test_html_head(self, fx_item_catalogue_min: ItemCatalogue):
+    def test_html_head(self, fx_item_catalogue_model_min: ItemCatalogue):
         """
         Can set common page elements in site layout.
 
         Integration test between item template and site layout.
         """
-        expected = fx_item_catalogue_min.page_metadata
-        html = BeautifulSoup(fx_item_catalogue_min.render(), parser="html.parser", features="lxml")
+        expected = fx_item_catalogue_model_min.page_metadata
+        html = BeautifulSoup(fx_item_catalogue_model_min.render(), parser="html.parser", features="lxml")
 
         assert html.head.title.string == expected.html_title